requirements.txt
streamlit
pytesseract
pillow
numpy
easyocr
